Move hazard "sets" out of the tooltip class and into hazard
